Reimplemented controller action reporting (WIP, tests pending)
For each action we now can tell (with acceptable accuracy) if it
comes from a gem, or if it comes from a module. Neither kind
of action is reported by default but they can be enabled with
the GEMS=1 and MODULES=1 arguments.

Source location for actions is by default shortened to a more
relevant portion, FULL_PATH=1 shows the full path.

In addition DUPLICATES=1 is what now enables listing actions inherited
from a superclass and METADATA=1 lists basic metadata per action.

ALL=1 makes the routes:missing task behave in self-contradictory way
by asking it to list every action instead of only focusing on those
that don't have any matching routes. This is convenient to get a
relatively compact sorted tree view of all the controllers and actions
in the application.